refactor(radio-group): merge into one file
