+++ packages/bezier-react/src/components/Forms/RadioGroup/index.ts
@@ -1,3 +1,3 @@
-export { RadioGroup } from './RadioGroup'
-export { Radio } from './Radio'
+export { Radio, RadioGroup } from './RadioGroup'
+
 export type { RadioGroupProps, RadioProps } from './RadioGroup.types'
